refactor(free-trial): separate free trial context and provider
